commiting to the product page revisoin
